requirements.txt
absl-py==2.4.0
aiofiles==25.1.0
anyio==4.12.1
argcomplete==3.6.3
astunparse==1.6.3
async-lru==2.2.0
attrs==25.4.0
backports.zstd==1.3.0
bcrypt==5.0.0
beautifulsoup4==4.14.3
boto3==1.42.65
botocore==1.42.65
brotli==1.2.0
cachetools==7.0.5
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
contourpy==1.3.3
cross-web==0.4.1
cryptography==46.0.5
cuda-bindings==12.9.4
cuda-pathfinder==1.4.1
cycler==0.12.1
dacite==1.9.2
Deprecated==1.3.1
dill==0.4.1
dnspython==2.8.0
exceptiongroup==1.3.1
fiftyone==1.13.4
fiftyone-brain==0.21.5
fiftyone_db==1.4.1
filelock==3.25.1
flatbuffers==25.12.19
fonttools==4.62.0
fsspec==2026.2.0
ftfy==6.3.1
gast==0.7.0
glob2==0.7
google-pasta==0.2.0
graphql-core==3.2.8
grpcio==1.78.0
h11==0.16.0
h2==4.3.0
h5py==3.16.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
humanize==4.15.0
Hypercorn==0.18.0
hyperframe==6.1.0
idna==3.11
ImageIO==2.37.3
inflate64==1.0.4
Jinja2==3.1.6
jmespath==1.1.0
joblib==1.5.3
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==3.0.0
kagglehub==1.0.0
kagglesdk==0.1.16
keras==3.13.2
keras-hub==0.26.0
kiwisolver==1.5.0
lazy-loader==0.5
libclang==18.1.1
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
ml_dtypes==0.5.4
mongoengine==0.29.3
motor==3.6.1
mpmath==1.3.0
multivolumefile==0.2.3
namex==0.1.0
narwhals==2.18.0
networkx==3.6.1
numpy==2.4.3
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
opencv-python==4.13.0.92
opencv-python-headless==4.13.0.92
opt_einsum==3.4.0
optree==0.19.0
packaging==26.0
pandas==3.0.1
paramiko==3.5.1
pillow==12.1.1
plotly==6.6.0
pprintpp==0.4.0
priority==2.0.0
protobuf==7.34.0
psutil==7.2.2
py7zr==1.1.0
pybcj==1.0.7
pycocotools==2.0.11
pycparser==3.0
pycryptodomex==3.23.0
pydash==8.0.6
Pygments==2.19.2
pymongo==4.9.2
PyNaCl==1.6.2
pyparsing==3.3.2
pyppmd==1.3.1
python-dateutil==2.9.0.post0
pytz==2026.1.post1
PyYAML==6.0.3
rarfile==4.2
regex==2026.2.28
requests==2.32.5
retrying==1.4.2
rich==14.3.3
rtree==1.4.1
s3transfer==0.16.0
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.17.1
setuptools==82.0.1
six==1.17.0
sortedcontainers==2.4.0
soupsieve==2.8.3
sse-starlette==0.10.3
sseclient-py==1.9.0
starlette==0.52.1
strawberry-graphql==0.291.3
sympy==1.14.0
tabulate==0.10.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
tensorflow-text==2.20.0
termcolor==3.3.0
texttable==1.7.0
threadpoolctl==3.6.0
tifffile==2026.3.3
torch==2.10.0
tqdm==4.67.3
triton==3.6.0
typing_extensions==4.15.0
tzlocal==5.3.1
universal-analytics-python3==1.1.1
urllib3==2.6.3
voxel51-eta==0.15.3
wcwidth==0.6.0
Werkzeug==3.1.6
wheel==0.46.3
wrapt==2.1.2
wsproto==1.3.2
xmltodict==1.0.4
